Retire old monitoring scripts and update workflow
